支持视频文件上传

+++ src/main/java/com/taotao/tool/dto/FileExt.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class FileExt {
 
     private String coverFilename;
-    private Integer second;
+    private Long second;
 }
